Detect disconnected device and avoid filetree corruption because of the cache
Use the "check_events" call to detect if the device was disconnected or
is not working anymore.
This is especially usefull for syscalls that are able to use the
gphotofs cache like getattr.
Without it, a client of the fuse can sometimes think that the device is
working but has an "empty" content or some "empty" folders when in fact
it is just that the device is in a bad state or removed.
So, this prevent a "copy" tool to report that a complete copy succeed
when it is infact not the case.

This needs the following patch of libgphoto2:
https://github.com/gphoto/libgphoto2/pull/189
(Otherwise, the behavior will be same as before)