Exit with a mount error if no or bad device detected
Move most of the "init" step inside the "main" function.
So, errors can be reported to the user at the mount stage and the fuse
mount be canceled early enough.
In the gphotofs_init(), fuse has already forked, and the mount is
already done, so it is too late to report any error.

Before this fix, it was possible to run gphotofs and have a mount point
even if there wasn't any device detected or if the connection parameters
were not good.

In addition, the init was not perfoming any real "request" to the device
if no particular option were used. So a lack of connected device was not
even detected.
A call to getstorage is done at the init stage to check that the device
is working and that it makes sense to let fuse create the mount point.

As in "simple-mtpfs", we exit with an error if no "storage" is detected.
It is the case for example with a "nexus" device that is in "charger"
mode.